Replaced variable type from double to long and added L as suffix to the long number to specify it as long instead of default integer

+++ src/Problem_3_LargestPrimeFactor.java
@@ -8,19 +8,19 @@ public class Problem_3_LargestPrimeFactor {
 
 
     public static void main (String[] args) {
-        Problem_3_LargestPrimeFactor findPrime = new Problem_3_LargestPrimeFactor(600851475143);
+        Problem_3_LargestPrimeFactor findPrime = new Problem_3_LargestPrimeFactor(600851475143L);
         findPrime.getAllFactors();
         System.out.println("The factors are: " + findPrime.factors );
         System.out.println("The max prime factor is: " + findPrime.maxPrimeFactor);
     }
 
 
-    private double number = 0;
-    private ArrayList<Double> factors = new ArrayList();
+    private long number = 0;
+    private ArrayList<Long> factors = new ArrayList();
     //private ArrayList primeFactors = new ArrayList();
-    private double maxPrimeFactor = 0;
+    private long maxPrimeFactor = 0;
 
-    public Problem_3_LargestPrimeFactor (double num) {
+    public Problem_3_LargestPrimeFactor (long num) {
         this.number = num;
 
     }
@@ -30,8 +30,8 @@ public class Problem_3_LargestPrimeFactor {
         getMaxPrimeFactor(this.factors);
     }
 
-    private void getFactors(double number) {
-        for (double i = 2; i <= number / 2; i++) {
+    private void getFactors(long number) {
+        for (long i = 2; i <= number / 2; i++) {
                 if (number % i ==0)
                     this.factors.add(number / i);
                     
@@ -39,11 +39,11 @@ public class Problem_3_LargestPrimeFactor {
     }
 
     //checks whether an int is prime or not.
-    private boolean isPrime(Double n) {
+    private boolean isPrime(long n) {
         //check if n is a multiple of 2
         if (n%2==0) return false;
         //if not, then just check the odds
-        for(double i=3;i*i<=n;i+=2) {
+        for(long i=3;i*i<=n;i+=2) {
             if(n%i==0)
                 return false;
         }
@@ -51,7 +51,7 @@ public class Problem_3_LargestPrimeFactor {
     }
 
     private void getMaxPrimeFactor(ArrayList factorsList) {
-        for (Double factor : this.factors
+        for (long factor : this.factors
              ) {
             if (isPrime(factor)) {
                 this.maxPrimeFactor = factor;
